complete

+++ resources/views/Comments/index.blade.php
@@ -5,35 +5,44 @@
     <div class="container-lg">
     <x-alert/>
     @auth
-        <form class="clearfix" id="comment_form" method="post" action="http://localhost/Vaii_sem_laravel/public/Comments/create">
+        <form class="clearfix pb-5  " id="comment_form" method="post" action="http://localhost/Vaii_sem_laravel/public/Comments/create">
             @csrf
-            <input type="hidden" name="createdBy" value="{{Auth::user()->id}}">
+            <input type="hidden" name="createdBy" value="{{Auth::user()->name}}">
             <textarea name="commentBody" id="comment_text" class="form-control" cols="30" rows="3"> </textarea>
             <button class="btn btn-primary btn-sm pull-right" name="submitComment" id="submitCommentID">Pridať recenziu</button>
         </form>
-    @elseauth()
-        <form class="clearfix" id="comment_form">
-            <p id="not_logged_in" cols="30" rows="10"> Pred pridaním recenzie sa musíte prihlásiť </p>
-        </form>
     @endauth
+    @guest()
+        <form class="clearfix" id="comment_form" >
+            <p id="not_logged_in" cols="30" rows="10" class="text-danger pb-5" > Pred pridaním recenzie sa musíte prihlásiť </p>
+        </form>
+    @endguest
 
 
     @foreach($comments as $comment)
-        <div class="comment clearfix" data-author="{{$comment->createdBy}}" id="comment{{$comment->id}}">
+        <div class="comment clearfix border-dark table-active rounded p-3 " data-author="{{$comment->createdBy}}" id="comment{{$comment->id}}">
             <div class="comment-details">
                 <p class="author" > {{$comment->createdBy}}  </p>
                 <span class="comment-date">{{$comment->created_at}}</span>
                 <p id="commentBodyID{{$comment->id}}" >{{$comment->commentBody}}</p>
-                <a  class="btn btn-warning commentDeleteBtn" onclick="deleteComment({{$comment->id}})" >Vymazať</a>
-                <a  class="btn btn-warning commentEditBtn"  onclick="editComment({{$comment->id}})" >Upraviť</a>
+                @auth
+                    @if((Auth::user()->name == 'Admin')||Auth::user()->name == $comment->createdBy)
+                        <a  class="btn btn-warning commentDeleteBtn" onclick="deleteComment({{$comment->id}})" >Vymazať</a>
+                        <a  class="btn btn-warning commentEditBtn"  onclick="editComment({{$comment->id}})" >Upraviť</a>
+                    @endif
+                @endauth
             </div>
         </div>
 
         <div class="comment clearfix edit_comment_form" id="editComment{{$comment->id}}">
             <div class="comment-details">
                 <textarea class="form-control" cols="10" rows="1" id="editCommentArea{{$comment->id}}">{{$comment->commentBody}}</textarea>
-                <a   class="btn btn-warning commentSendEditBtn" onclick="updateComment({{$comment->id}})" >Odoslať</a>
-                <a   class="btn btn-warning commentCancelEditBtn" onclick="cancelEdit({{$comment->id}})" >Zrusit</a>
+                @auth
+                    @if((Auth::user()->name == 'Admin')||Auth::user()->name == $comment->createdBy)
+                        <a   class="btn btn-warning commentSendEditBtn" onclick="updateComment({{$comment->id}})" >Odoslať</a>
+                        <a   class="btn btn-warning commentCancelEditBtn" onclick="cancelEdit({{$comment->id}})" >Zrusit</a>
+                    @endif
+                @endauth
             </div>
         </div>
